pyproject.toml
[project]
name = "mcp-client-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agno>=2.1.0",
    "fastapi>=0.100.0",
    "httpx>=0.25.0",
    "mcp>=1.0.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "sqlalchemy>=2.0.0",
    "typer>=0.9.0",
    "uvicorn>=0.20.0",
]
